Added todo
Former-commit-id: 2a4182f81d3e8cf10cc054a202a3369d83f16d3c [formerly e963e4e725c140db6a77261bddb85284b8266027]
Former-commit-id: 8dce4bb6b9ca3f0312d5eb9154751f3de4048c24 [formerly 3852e16ccb7ec258803a0810fcca500864e233f9]
Former-commit-id: d72b4231c29fba86840f990f1a08cd96271792a3
Former-commit-id: 3162c71c545fa342d2621bf7d49b48d2185678e1

+++ Source/RawParserUWP/RawParserUWP/Model/Format/Image/Base/RawImage.cs
@@ -77,14 +77,14 @@ namespace RawParser.Model.ImageDisplay
         /*
          * For testing
          */
-        public byte[] getImageAsByteArray()
+        public ushort[] getImageAsByteArray()
         {
-            byte[] tempByteArray = new byte[width * height];
+            ushort[] tempByteArray = new ushort[width * height];
 
             for (int i = 0; i < width * height; i++)
             {
                 //get the pixel
-                byte temp = 0;
+                ushort temp = 0;
 
                 for (int k = 0; k < 8; k++)
                 {
@@ -93,7 +93,7 @@ namespace RawParser.Model.ImageDisplay
                     //if(xy)Debug.WriteLine("value: " + xy);
                     if (xy)
                     {
-                        temp |= (byte)(1 << k);
+                        temp |= (ushort)(1 << k);
                     }
                 }
